+++ lib/domains/performance.dart
@@ -2,10 +2,10 @@ import 'dart:async';
 import 'package:meta/meta.dart' show required;
 import '../src/connection.dart';
 
-class PerformanceManager {
+class PerformanceApi {
   final Client _client;
 
-  PerformanceManager(this._client);
+  PerformanceApi(this._client);
 
   /// Current values of the metrics.
   Stream<MetricsEvent> get onMetrics => _client.onEvent