+++ lib/src/connection.dart
@@ -24,13 +24,13 @@ class Connection implements Client {
   final List<Session> _sessions = [];
   final StreamController<Event> _eventController =
       new StreamController<Event>.broadcast();
-  TargetManager _targets;
+  TargetApi _targets;
   final List<StreamSubscription> _subscriptions = [];
 
   Connection._(this._webSocket) {
     _subscriptions.add(_webSocket.listen(_onMessage));
 
-    _targets = new TargetManager(this);
+    _targets = new TargetApi(this);
 
     _subscriptions.add(_targets.onReceivedMessageFromTarget
         .listen((ReceivedMessageFromTargetEvent e) {
@@ -45,7 +45,7 @@ class Connection implements Client {
     }));
   }
 
-  TargetManager get targets => _targets;
+  TargetApi get targets => _targets;
 
   Session _getSession(SessionID sessionId) =>
       _sessions.firstWhere((s) => s.sessionId.value == sessionId.value);
@@ -139,13 +139,13 @@ class Session implements Client {
   static int _lastId = 0;
   final TargetID targetID;
   final SessionID sessionId;
-  final TargetManager _targetManager;
+  final TargetApi _targetApi;
   final BrowserContextID _browserContextID;
   final Map<int, Completer> _completers = {};
   final StreamController<Event> _eventController =
       new StreamController<Event>.broadcast();
 
-  Session._(this._targetManager, this.targetID, this.sessionId,
+  Session._(this._targetApi, this.targetID, this.sessionId,
       {BrowserContextID browserContextID})
       : _browserContextID = browserContextID;
 
@@ -160,7 +160,7 @@ class Session implements Client {
     Completer completer = new Completer();
     _completers[id] = completer;
 
-    _targetManager.sendMessageToTarget(message, sessionId: sessionId);
+    _targetApi.sendMessageToTarget(message, sessionId: sessionId);
 
     return completer.future;
   }
@@ -194,9 +194,9 @@ class Session implements Client {
   }
 
   Future close() async {
-    await _targetManager.closeTarget(targetID);
+    await _targetApi.closeTarget(targetID);
     if (_browserContextID != null) {
-      await _targetManager.disposeBrowserContext(_browserContextID);
+      await _targetApi.disposeBrowserContext(_browserContextID);
     }
   }
 }