+++ lib/src/wait_until.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
-import '../domains/network.dart';
 
-Future waitUntilNetworkIdle(NetworkManager network,
+import '../domains/network.dart';
+import '../domains/log.dart';
+
+Future waitUntilNetworkIdle(NetworkApi network,
     {Duration idleDuration: const Duration(milliseconds: 1000),
     int idleInFlight: 0}) async {
   await network.enable();
@@ -43,4 +45,13 @@ Future waitUntilNetworkIdle(NetworkManager network,
   }));
 
   await completer.future;
+}
+
+Future waitUntilConsoleContains(LogApi log, String text) async {
+  await log.enable();
+  await for (LogEntry logEntry in log.onEntryAdded) {
+    if (logEntry.text.contains(text)) {
+      return;
+    }
+  }
 }