+++ tool/generate_readme.dart
@@ -32,5 +32,7 @@ String generateReadme() {
         '[$protocolName]()', '[$protocolName](${protocols[protocolName]})');
   }
 
+  //TODO(xha): generate TableOfContent
+
   return readme;
 }